# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: UserSrv Service
  description: HTTP service for managing user-srv service
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /file/upload:
    post:
      tags:
      - file
      summary: upload file
      description: Upload static file
      operationId: file#upload
      parameters:
      - name: UploadRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/FileUploadRequestBody'
          required:
          - file
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/FileUploadResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/Fileupload_file_upload_err_response_body'
      schemes:
      - http
  /swagger/swagger.json:
    get:
      summary: Download gen/http/openapi.json
      description: JSON document containing the API swagger definition
      operationId: swagger#/swagger/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
  /user/current_user:
    get:
      tags:
      - user
      summary: retrieve user
      description: Show user info by Token
      operationId: user#retrieve
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserRetrieveResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/Userretrieve_invalide_token_response_body'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/Userretrieve_not_found_response_body'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /user/email:
    post:
      tags:
      - user
      summary: sendEmail user
      description: Send email to active user
      operationId: user#sendEmail
      parameters:
      - name: SendEmailRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserSendEmailRequestBody'
          required:
          - email
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserSendEmailResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/UsersendEmail_email_not_found_response_body'
      schemes:
      - http
  /user/email/{code}:
    get:
      tags:
      - user
      summary: activeEmail user
      description: Active email to user
      operationId: user#activeEmail
      parameters:
      - name: code
        in: path
        description: The code for email to active
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserActiveEmailResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/UseractiveEmail_code_invalide_response_body'
      schemes:
      - http
  /user/register:
    post:
      tags:
      - user
      summary: create user
      description: Add new user
      operationId: user#create
      parameters:
      - name: CreateRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserCreateRequestBody'
          required:
          - email
          - password
      responses:
        "201":
          description: Created response.
          schema:
            $ref: '#/definitions/UserCreateResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/Usercreate_email_exist_response_body'
      schemes:
      - http
  /user/signin:
    post:
      tags:
      - user
      summary: signin user
      description: Creates a valid JWT
      operationId: user#signin
      parameters:
      - name: SigninRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserSigninRequestBody'
          required:
          - email
          - password
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserSigninResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/Usersignin_email_or_password_error_response_body'
      schemes:
      - http
  /user/update:
    put:
      tags:
      - user
      summary: update user
      description: Update avatar and nickname about user
      operationId: user#update
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: true
        type: string
      - name: UpdateRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserUpdateRequestBody'
          required:
          - avatar
          - nickname
      responses:
        "204":
          description: No Content response.
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/Userupdate_invalide_token_response_body'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/Userupdate_not_found_response_body'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
definitions:
  FileUploadRequestBody:
    title: FileUploadRequestBody
    type: object
    properties:
      file:
        type: string
        description: file
        example: Facilis tenetur.
    example:
      file: Dolor aliquid blanditiis fugiat quae.
    required:
    - file
  FileUploadResponseBody:
    title: 'Mediatype identifier: responsedata; view=default'
    type: object
    properties:
      code:
        type: integer
        description: code
        default: 200
        example: 200
        format: int64
      data:
        type: string
        example: Animi quos error libero et.
      message:
        type: string
        description: message
        default: success
        example: success
    description: UploadResponseBody result type (default view)
    example:
      code: 200
      data: Nostrum velit eaque molestias sunt voluptatem.
      message: success
    required:
    - code
    - message
    - data
  Fileupload_file_upload_err_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: file upload error (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UserActiveEmailResponseBody:
    title: 'Mediatype identifier: responsedata; view=default'
    type: object
    properties:
      code:
        type: integer
        description: code
        default: 200
        example: 200
        format: int64
      data:
        type: string
        example: Et deleniti officiis molestiae modi.
      message:
        type: string
        description: message
        default: success
        example: success
    description: ActiveEmailResponseBody result type (default view)
    example:
      code: 200
      data: Quam ullam sed alias vitae ullam.
      message: success
    required:
    - code
    - message
    - data
  UserCreateRequestBody:
    title: UserCreateRequestBody
    type: object
    properties:
      email:
        type: string
        description: email of userProfile
        example: 123@456.com
        pattern: \S+@\S+\.\S+
      password:
        type: string
        description: password of userProfile
        example: "123456"
        minLength: 6
    example:
      email: 123@456.com
      password: "123456"
    required:
    - email
    - password
  UserCreateResponseBody:
    title: 'Mediatype identifier: responsedata; view=default'
    type: object
    properties:
      code:
        type: integer
        description: code
        default: 200
        example: 200
        format: int64
      data:
        type: string
        example: Quia tenetur minima aliquid unde dicta aliquid.
      message:
        type: string
        description: message
        default: success
        example: success
    description: CreateResponseBody result type (default view)
    example:
      code: 200
      data: Aliquam odio.
      message: success
    required:
    - code
    - message
    - data
  UserRetrieveResponseBody:
    title: 'Mediatype identifier: responsedata; view=default'
    type: object
    properties:
      code:
        type: integer
        description: code
        default: 200
        example: 200
        format: int64
      data:
        type: string
        example: Rem omnis corrupti qui voluptatum et vel.
      message:
        type: string
        description: message
        default: success
        example: success
    description: RetrieveResponseBody result type (default view)
    example:
      code: 200
      data: Voluptatum in est blanditiis qui.
      message: success
    required:
    - code
    - message
    - data
  UserSendEmailRequestBody:
    title: UserSendEmailRequestBody
    type: object
    properties:
      email:
        type: string
        description: email of userProfile
        example: 123@456.com
        pattern: \S+@\S+\.\S+
    example:
      email: 123@456.com
    required:
    - email
  UserSendEmailResponseBody:
    title: 'Mediatype identifier: responsedata; view=default'
    type: object
    properties:
      code:
        type: integer
        description: code
        default: 200
        example: 200
        format: int64
      data:
        type: string
        example: Placeat ab.
      message:
        type: string
        description: message
        default: success
        example: success
    description: SendEmailResponseBody result type (default view)
    example:
      code: 200
      data: Sequi est rem adipisci id amet consequatur.
      message: success
    required:
    - code
    - message
    - data
  UserSigninRequestBody:
    title: UserSigninRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email used to perform signin
        example: 1@1.com
      password:
        type: string
        description: Password used to perform signin
        example: "123456"
    example:
      email: 1@1.com
      password: "123456"
    required:
    - email
    - password
  UserSigninResponseBody:
    title: 'Mediatype identifier: responsedata; view=default'
    type: object
    properties:
      code:
        type: integer
        description: code
        default: 200
        example: 200
        format: int64
      data:
        type: string
        example: Sit consectetur sunt.
      message:
        type: string
        description: message
        default: success
        example: success
    description: SigninResponseBody result type (default view)
    example:
      code: 200
      data: Aut ut voluptatem non accusantium rerum.
      message: success
    required:
    - code
    - message
    - data
  UserUpdateRequestBody:
    title: UserUpdateRequestBody
    type: object
    properties:
      avatar:
        type: string
        description: avatar of user
        example: https://www.baidu.com/img/bd_logo1.png?where=super
      nickname:
        type: string
        description: nickname of user
        example: Bobby
    example:
      avatar: https://www.baidu.com/img/bd_logo1.png?where=super
      nickname: Bobby
    required:
    - avatar
    - nickname
  UseractiveEmail_code_invalide_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: code_invalide (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  Usercreate_email_exist_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: email exist (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  Userretrieve_invalide_token_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: invalide token (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  Userretrieve_not_found_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: User not found (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  UsersendEmail_email_not_found_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: email not found (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  Usersignin_email_or_password_error_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: email or password error (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  Userupdate_invalide_token_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: invalide token (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  Userupdate_not_found_response_body:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: user not found (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: Secures endpoint by requiring a valid JWT token retrieved via the
      signin endpoint.
    name: Authorization
    in: header